+++ src/commands/clean.ts
@@ -29,8 +29,10 @@ export default class Clean extends Command {
       }
 
       if (Utils.isRoot()) {
-         const bleach = new Bleach()
-         bleach.clean(verbose)
+         if (await await Utils.customConfirm(`Select yes to continue...`)) {
+            const bleach = new Bleach()
+            bleach.clean(verbose)
+         }
       }
    }
 }